vochain: refactor transaction check
Unify all transaction check methods.
