+++ vochain/state/account.go
@@ -190,81 +190,6 @@ func (v *State) CreateAccount(accountAddress common.Address, infoURI string, del
 	return v.SetAccount(accountAddress, newAccount)
 }
 
-// SetAccountInfoTxCheck checks if a set account info tx is valid
-func SetAccountInfoTxCheck(vtx *models.Tx, txBytes, signature []byte, state *State) error {
-	if vtx == nil || signature == nil || txBytes == nil || state == nil {
-		return fmt.Errorf("invalid transaction parameters provided")
-	}
-	tx := vtx.GetSetAccount()
-	if tx == nil {
-		return fmt.Errorf("invalid transaction")
-	}
-	pubKey, err := ethereum.PubKeyFromSignature(txBytes, signature)
-	if err != nil {
-		return fmt.Errorf("cannot extract public key from signature: %w", err)
-	}
-	txSenderAddress, err := ethereum.AddrFromPublicKey(pubKey)
-	if err != nil {
-		return fmt.Errorf("cannot extract address from public key: %w", err)
-	}
-	txAccountAddress := common.BytesToAddress(tx.GetAccount())
-	if txAccountAddress == (common.Address{}) {
-		txAccountAddress = txSenderAddress
-	}
-	txSenderAccount, err := state.GetAccount(txSenderAddress, false)
-	if err != nil {
-		return fmt.Errorf("cannot check if account %s exists: %w", txSenderAddress, err)
-	}
-	if txSenderAccount == nil {
-		return ErrAccountNotExist
-	}
-	// check txSender nonce
-	if tx.GetNonce() != txSenderAccount.Nonce {
-		return fmt.Errorf(
-			"invalid nonce, expected %d got %d",
-			txSenderAccount.Nonce,
-			tx.GetNonce(),
-		)
-	}
-	// get setAccount tx cost
-	costSetAccountInfoURI, err := state.TxCost(models.TxType_SET_ACCOUNT_INFO_URI, false)
-	if err != nil {
-		return fmt.Errorf("cannot get tx cost: %w", err)
-	}
-	// check tx sender balance
-	if txSenderAccount.Balance < costSetAccountInfoURI {
-		return fmt.Errorf("unauthorized: %s", ErrNotEnoughBalance)
-	}
-	// check info URI
-	infoURI := tx.GetInfoURI()
-	if len(infoURI) == 0 || len(infoURI) > types.MaxURLLength {
-		return fmt.Errorf("invalid URI, cannot be empty")
-	}
-	if txSenderAddress == txAccountAddress {
-		if infoURI == txSenderAccount.InfoURI {
-			return fmt.Errorf("invalid URI, must be different")
-		}
-		return nil
-	}
-	// if txSender != txAccount only delegate operations
-	// get tx account Account
-	txAccountAccount, err := state.GetAccount(txAccountAddress, false)
-	if err != nil {
-		return fmt.Errorf("cannot get tx account: %w", err)
-	}
-	if txAccountAccount == nil {
-		return ErrAccountNotExist
-	}
-	if infoURI == txAccountAccount.InfoURI {
-		return fmt.Errorf("invalid URI, must be different")
-	}
-	// check if delegate
-	if !txAccountAccount.IsDelegate(txSenderAddress) {
-		return fmt.Errorf("tx sender is not a delegate")
-	}
-	return nil
-}
-
 // SetAccount sets the given account data to the state
 func (v *State) SetAccount(accountAddress common.Address, account *Account) error {
 	accBytes, err := proto.Marshal(account)
